# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: OneChat API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /matches/{matchId}/messages/{messageId}/reactions:
    get:
      operationId: getMessageReactions
      tags: [reactions]
      summary: Get reactions for a message
      parameters:
        - name: matchId
          in: path
          required: true
          schema:
            type: string
        - name: messageId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Reaction summary
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReactionSummary"
    post:
      operationId: addReaction
      tags: [reactions]
      summary: Add a reaction to a message
      parameters:
        - name: matchId
          in: path
          required: true
          schema:
            type: string
        - name: messageId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReactionInput"
      responses:
        "201":
          description: Reaction added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReactionAdded"
  /matches/{matchId}/messages/{messageId}/reactions/{emoji}:
    delete:
      operationId: removeReaction
      tags: [reactions]
      summary: Remove a reaction from a message
      parameters:
        - name: matchId
          in: path
          required: true
          schema:
            type: string
        - name: messageId
          in: path
          required: true
          schema:
            type: string
        - name: emoji
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Reaction removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /reports:
    post:
      operationId: createReport
      tags: [moderation]
      summary: Report a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportInput"
      responses:
        "201":
          description: Report submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportCreated"
...
  /blocks/{blockedId}:
    delete:
      operationId: unblockUser
      tags: [blocks]
      summary: Unblock a user
      parameters:
        - name: blockedId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User unblocked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
...
  schemas:
...
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
...
    ReactionSummary:
      type: object
      required: [emoji, count, byMe]
      properties:
        emoji:
          type: string
        count:
          type: integer
        byMe:
          type: boolean
    ReactionInput:
      type: object
      required: [emoji]
      properties:
        emoji:
          type: string
    ReactionAdded:
      type: object
      required: [messageId, emoji, userId]
      properties:
        messageId:
          type: string
        emoji:
          type: string
        userId:
          type: string
    ReportInput:
      type: object
      required: [reportedUserId, reason]
      properties:
        reportedUserId:
          type: string
        matchId:
          type: ["string", "null"]
        reason:
          type: string
          enum: [spam, harassment, inappropriate, other]
        details:
          type: ["string", "null"]
          maxLength: 1000
    ReportCreated:
      type: object
      required: [id, success]
      properties:
        id:
          type: string
        success:
          type: boolean
...