# lib/api-zod/src/generated/types/reactionAdded.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OneChat API specification
 * OpenAPI spec version: 0.2.0
 */

export interface ReactionAdded {
  messageId: string;
  emoji: string;
  userId: string;
}