# lib/api-zod/src/generated/types/reportCreated.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OneChat API specification
 * OpenAPI spec version: 0.2.0
 */

export interface ReportCreated {
  id: string;
  success: boolean;
}