# lib/api-zod/src/generated/types/reportInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OneChat API specification
 * OpenAPI spec version: 0.2.0
 */
import type { ReportInputReason } from './reportInputReason';

export interface ReportInput {
  reportedUserId: string;
  /** @nullable */
  matchId?: string | null;
  reason: ReportInputReason;
  /**
     * @maxLength 1000
     * @nullable
     */
  details?: string | null;
}